ref(storage): lift alternate root and multiplexing metastores
Refactors storage to wrap existing underlying notary stores. This
simplifies the storage layer and will more easily support non-memory
backends.
