+++ client.Dockerfile
@@ -15,7 +15,10 @@ RUN apt-get update && apt-get install -y \
 
 RUN useradd -ms /bin/bash notary \
 	&& pip install codecov \
-	&& go get github.com/golang/lint/golint github.com/fzipp/gocyclo github.com/client9/misspell/cmd/misspell github.com/gordonklaus/ineffassign github.com/HewlettPackard/gas
+	&& go get github.com/golang/lint/golint github.com/fzipp/gocyclo \
+	   github.com/client9/misspell/cmd/misspell \
+	   github.com/gordonklaus/ineffassign \
+	   github.com/HewlettPackard/gas
 
 RUN go get github.com/docker/notary
